refactoring for mvp
