+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.main.MainActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
@@ -25,22 +25,22 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ExplanationActivity"
+            android:name=".ui.explanation.ExplanationActivity"
             android:label="Answer and Explanation"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName=".ui.main.MainActivity"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.vsklamm.cppquiz.MainActivity" />
+                android:value="com.vsklamm.cppquiz.ui.main.MainActivity" />
         </activity>
         <activity
-            android:name=".AboutActivity"
+            android:name=".ui.about.AboutActivity"
             android:label="@string/about"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName=".ui.main.MainActivity"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.vsklamm.cppquiz.MainActivity" />
+                android:value="com.vsklamm.cppquiz.ui.main.MainActivity" />
         </activity>
     </application>
