+++ app/src/main/java/com/vsklamm/cppquiz/data/database/AppDatabase.java
@@ -1,9 +1,9 @@
-package com.vsklamm.cppquiz.database;
+package com.vsklamm.cppquiz.data.database;
 
 import android.arch.persistence.room.Database;
 import android.arch.persistence.room.RoomDatabase;
 
-import com.vsklamm.cppquiz.model.Question;
+import com.vsklamm.cppquiz.data.Question;
 
 @Database(entities = {Question.class}, version = 1, exportSchema = false)
 public abstract class AppDatabase extends RoomDatabase {
