+++ app/src/main/java/com/vsklamm/cppquiz/data/database/QuestionDao.java
@@ -1,4 +1,4 @@
-package com.vsklamm.cppquiz.database;
+package com.vsklamm.cppquiz.data.database;
 
 import android.arch.persistence.room.Dao;
 import android.arch.persistence.room.Delete;
@@ -7,7 +7,7 @@ import android.arch.persistence.room.OnConflictStrategy;
 import android.arch.persistence.room.Query;
 import android.arch.persistence.room.Update;
 
-import com.vsklamm.cppquiz.model.Question;
+import com.vsklamm.cppquiz.data.Question;
 
 import java.util.List;
