--- app/src/main/java/com/vsklamm/cppquiz/loader/BadResponseException.java
@@ -1,16 +1,0 @@
-package com.vsklamm.cppquiz.loader;
-
-public class BadResponseException extends Exception {
-
-    public BadResponseException(String message) {
-        super(message);
-    }
-
-    public BadResponseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public BadResponseException(Throwable cause) {
-        super(cause);
-    }
-}
