+++ app/src/main/java/com/vsklamm/cppquiz/loader/DumpLoader.java
@@ -8,12 +8,12 @@ import android.support.annotation.Nullable;
 import android.support.v4.content.AsyncTaskLoader;
 
 import com.vsklamm.cppquiz.App;
-import com.vsklamm.cppquiz.MainActivity;
-import com.vsklamm.cppquiz.api.CppQuizLiteApi;
-import com.vsklamm.cppquiz.database.AppDatabase;
-import com.vsklamm.cppquiz.database.QuestionDao;
+import com.vsklamm.cppquiz.ui.main.MainActivity;
+import com.vsklamm.cppquiz.data.api.CppQuizLiteApi;
+import com.vsklamm.cppquiz.data.database.AppDatabase;
+import com.vsklamm.cppquiz.data.database.QuestionDao;
 import com.vsklamm.cppquiz.model.DumpDataType;
-import com.vsklamm.cppquiz.model.Question;
+import com.vsklamm.cppquiz.data.Question;
 import com.vsklamm.cppquiz.utils.Parser;
 import com.vsklamm.cppquiz.utils.RequestType;
 import com.vsklamm.cppquiz.utils.TimeWork;
