--- app/src/main/java/com/vsklamm/cppquiz/quiz/MainView.java
@@ -1,5 +1,0 @@
-package com.vsklamm.cppquiz.quiz;
-
-public interface  MainView {
-
-}
