+++ app/src/main/java/com/vsklamm/cppquiz/ui/BaseModeContract.java
@@ -1,5 +1,8 @@
-package com.vsklamm.cppquiz.quiz;
+package com.vsklamm.cppquiz.ui;
 
+/**
+ * This specifies the contract between the view and the presenter.
+ */
 public interface BaseModeContract {
 
     interface View extends BaseView<Presenter> {
