+++ app/src/main/java/com/vsklamm/cppquiz/ui/about/AboutActivity.java
@@ -1,8 +1,10 @@
-package com.vsklamm.cppquiz;
+package com.vsklamm.cppquiz.ui.about;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.MenuItem;
+
+import com.vsklamm.cppquiz.R;
 
 public class AboutActivity extends AppCompatActivity {
