+++ app/src/main/java/com/vsklamm/cppquiz/ui/dialogs/ConfirmHintDialog.java
@@ -1,4 +1,4 @@
-package com.vsklamm.cppquiz;
+package com.vsklamm.cppquiz.ui.dialogs;
 
 import android.app.Dialog;
 import android.content.Context;
@@ -8,7 +8,9 @@ import android.support.annotation.NonNull;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatDialogFragment;
 
-public class SelectQuizDialog extends AppCompatDialogFragment {
+import com.vsklamm.cppquiz.R;
+
+public class ConfirmHintDialog extends AppCompatDialogFragment {
 
     private DialogListener listener;
 
@@ -18,7 +20,8 @@ public class SelectQuizDialog extends AppCompatDialogFragment {
         AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
 
         alertDialogBuilder
-                .setTitle(R.string.select_quiz_title)
+                .setTitle(R.string.dialog_hint_title)
+                .setMessage(R.string.dialog_hint_text)
                 .setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
@@ -28,7 +31,7 @@ public class SelectQuizDialog extends AppCompatDialogFragment {
                 .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                        listener.onQuizSelected();
+                        listener.onConfirmedHintLoad();
                     }
                 });
         return alertDialogBuilder.create();
@@ -45,6 +48,6 @@ public class SelectQuizDialog extends AppCompatDialogFragment {
     }
 
     public interface DialogListener {
-        void onQuizSelected();
+        void onConfirmedHintLoad();
     }
 }
