+++ app/src/main/java/com/vsklamm/cppquiz/ui/main/GameLogic.java
@@ -1,12 +1,13 @@
-package com.vsklamm.cppquiz.quiz;
+package com.vsklamm.cppquiz.ui.main;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.util.SparseIntArray;
 
 import com.vsklamm.cppquiz.App;
-import com.vsklamm.cppquiz.database.AppDatabase;
-import com.vsklamm.cppquiz.model.Question;
+import com.vsklamm.cppquiz.data.UserData;
+import com.vsklamm.cppquiz.data.database.AppDatabase;
+import com.vsklamm.cppquiz.data.Question;
 
 import java.io.Serializable;
 import java.lang.ref.WeakReference;
