+++ app/src/main/java/com/vsklamm/cppquiz/ui/main/MainView.java
@@ -1,0 +1,5 @@
+package com.vsklamm.cppquiz.ui.main;
+
+public interface  MainView {
+
+}
