+++ app/src/main/java/com/vsklamm/cppquiz/ui/main/TrainingModeContract.java
@@ -1,0 +1,16 @@
+package com.vsklamm.cppquiz.ui.main;
+
+import com.vsklamm.cppquiz.ui.BasePresenter;
+import com.vsklamm.cppquiz.ui.BaseView;
+
+public interface TrainingModeContract {
+
+    interface View extends BaseView<Presenter> {
+
+    }
+
+    interface Presenter extends BasePresenter {
+
+    }
+
+}
