+++ app/src/main/res/layout/activity_explanation.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/style_triangular_halves"
-    tools:context=".ExplanationActivity">
+    tools:context=".ui.explanation.ExplanationActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
