+++ app/src/main/res/values/colors.xml
@@ -19,6 +19,6 @@
     <color name="hint_exp_title">#FFD600</color>
     <color name="explanation_correct">#B2DFDB</color>
 
-    <color name="shackbar_action_retry">#f06292</color>
+    <color name="shackbar_action_retry">#feb650</color>
 
 </resources>
